Split the startup extension into their respective projects as discussed in the PR110
Fixed the startup code after my initialization refactor
