+++ src/Sidekick.Core/StartupExtensions.cs
@@ -1,0 +1,17 @@
+using Microsoft.Extensions.DependencyInjection;
+using Sidekick.Core.Initialization;
+using Sidekick.Core.Loggers;
+
+namespace Sidekick.Core
+{
+    public static class StartupExtensions
+    {
+        public static IServiceCollection AddSidekickCoreServices(this IServiceCollection services)
+        {
+            services.AddSingleton<ILogger, Logger>();
+            services.AddSingleton<IInitializeService, InitializeService>();
+
+            return services;
+        }
+    }
+}
